# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API",
        "description": "Documentation for v1 api of pet store , this API is a recruitement test from Buckhill.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "View user account",
                "operationId": "267147be6383186a359317a4a3425848",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0IiwiYXVkIjoiaHR0cDovL2xvY2FsaG9zdCIsInN1YiI........."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string|null",
                                            "example": null
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "extra": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unproccessable content"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user account",
                "operationId": "c4c9cdd163b524d4cf8cbd0ec913e4f1",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user/orders": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List all orders for the authenticated user",
                "operationId": "48915a084884ffa876b1876c1c513311",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user/create": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create new user account",
                "operationId": "69202a1683555f6d82bf26ea4dd57cfe",
                "responses": {
                    "200": {
                        "description": "New user created with success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user/forgot-password": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a token to reset user password",
                "operationId": "9392cd6eb121838e272437f5012995b5",
                "responses": {
                    "200": {
                        "description": "New user created with success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unproccessable content"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login to his account",
                "operationId": "aaf752a8ae4cc5fdc238216e60863af5",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "incorrect email/password"
                    },
                    "404": {
                        "description": "Route not found"
                    },
                    "422": {
                        "description": "Unproccessable content"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Logout user",
                "operationId": "e68c6e0f387fbdc7c4d8220ac3f0b2f9",
                "responses": {
                    "200": {
                        "description": "Logout successful"
                    },
                    "404": {
                        "description": "Route not found"
                    },
                    "422": {
                        "description": "Unproccessable content"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user/reset-password-token": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Reset user password with token",
                "operationId": "e5d2435d136e0e16f508018cacec35a7",
                "responses": {
                    "200": {
                        "description": "New user created with success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user/edit": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update use account",
                "operationId": "1571f9a15c769b83196e134b74096d21",
                "responses": {
                    "200": {
                        "description": "New user created with success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Route not found"
                    },
                    "422": {
                        "description": "Unproccessable content"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: 'Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}